Nuova rete

+++ data_utils.py
@@ -17,21 +17,18 @@ def prepare_dataloaders(X, y, class_names,
                         random_state=42):
     num_classes = len(class_names)
     X_tr, X_te, y_tr, y_te = train_test_split(
-        X, y,
-        test_size=test_size,
-        random_state=random_state,
-        shuffle=True
+        X, y, test_size=test_size, random_state=random_state, shuffle=True
     )
-    # tensori e one-hot
+    # to torch tensors + reshape
     X_tr = torch.tensor(X_tr, dtype=torch.float32).permute(0,3,1,2)
     X_te = torch.tensor(X_te, dtype=torch.float32).permute(0,3,1,2)
-    Y_tr = F.one_hot(torch.tensor(y_tr), num_classes).float()
-    Y_te = F.one_hot(torch.tensor(y_te), num_classes).float()
+    Y_tr = F.one_hot(torch.tensor(y_tr, dtype=torch.long), num_classes).float()
+    Y_te = F.one_hot(torch.tensor(y_te, dtype=torch.long), num_classes).float()
 
     train_ds = TensorDataset(X_tr, Y_tr)
     test_ds  = TensorDataset(X_te, Y_te)
-    train_loader = DataLoader(train_ds, batch_size=batch_size, shuffle=True,
-                              num_workers=4, pin_memory=True)
-    test_loader  = DataLoader(test_ds,  batch_size=batch_size, shuffle=False,
-                              num_workers=4, pin_memory=True)
+    train_loader = DataLoader(train_ds, batch_size=batch_size,
+                              shuffle=True,  num_workers=4, pin_memory=True)
+    test_loader  = DataLoader(test_ds,  batch_size=batch_size,
+                              shuffle=False, num_workers=4, pin_memory=True)
     return train_loader, test_loader, num_classes
